chatBot-backend/requirements.txt
fastapi
uvicorn
motor
bcrypt
python-dotenv
openai